feat(solicitudes): solving issues
