+++ packages/solicitudes/src/components/utils/submitNewSolicitud.js
@@ -1,21 +1,18 @@
-export default function submitNewSolicitud() {
-/* solicitudErrors,
+export default function submitNewSolicitud(
+  solicitudErrors,
   error,
   form,
-  setNoti, */
-  console.log('nueva solicitud');
-  /* if (
+  setNoti,
+) {
+  if (
     Object.values(solicitudErrors).every((validation) => validation()) !== false
   ) {
     if (Object.values(error).every((value) => value === null || value === '')) {
-      fetch(
-        'http://localhost:3000/api/v1/solicitudes/',
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(form),
-        },
-      );
+      fetch('http://localhost:3000/api/v1/solicitudes', {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(form),
+      });
     } else {
       setNoti({
         open: true,
@@ -29,5 +26,5 @@ export default function submitNewSolicitud() {
       message: 'Algo salio mal, revisa que los campos esten correctos',
       type: 'error',
     });
-  } */
+  }
 }
